refactor ssh fixtures
